[FrontEnd]Tweaks to Register Requests

+++ src/main/java/io/cristaling/iss/reddrop/web/controllers/BloodRequestController.java
@@ -6,6 +6,8 @@ import io.cristaling.iss.reddrop.core.BloodRequest;
 import io.cristaling.iss.reddrop.core.Doctor;
 import io.cristaling.iss.reddrop.repositories.BloodRequestRepository;
 import io.cristaling.iss.reddrop.services.BloodRequestService;
+import io.cristaling.iss.reddrop.services.PermissionsService;
+import io.cristaling.iss.reddrop.utils.Permission;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
@@ -19,32 +21,45 @@ import java.util.UUID;
 public class BloodRequestController {
 
     BloodRequestService bloodRequestService;
+    PermissionsService permissionsService;
 
     @Autowired
-    public BloodRequestController(BloodRequestService bloodRequestService) {
+    public BloodRequestController(BloodRequestService bloodRequestService, PermissionsService permissionsService) {
         this.bloodRequestService = bloodRequestService;
+        this.permissionsService = permissionsService;
     }
 
     @RequestMapping("/add")
     @ResponseStatus(value = HttpStatus.NO_CONTENT)
-    public void addRequest(@RequestBody BloodRequest bloodRequest){
+    public void addRequest(UUID token, @RequestBody BloodRequest bloodRequest) {
+        if (!permissionsService.hasPermission(token, Permission.ADMIN)) {
+            return;
+        }
         bloodRequestService.save(bloodRequest);
     }
 
     @RequestMapping("/delete")
     @ResponseStatus(value = HttpStatus.NO_CONTENT)
-    public void deleteRequest(String uuid){
+    public void deleteRequest(UUID token, String uuid) {
+        if (!permissionsService.hasPermission(token, Permission.ADMIN)) {
+            return;
+        }
         bloodRequestService.deleteById(uuid);
     }
 
-
     @RequestMapping("/getall")
-    public List<BloodRequest> getAllRequests(){
+    public List<BloodRequest> getAllRequests(UUID token) {
+        if (!permissionsService.hasPermission(token, Permission.ADMIN)) {
+            return null;
+        }
         return bloodRequestService.findAll();
     }
 
     @RequestMapping("/getfrom")
-    public List<BloodRequest> getSpecificRequests(String uuid){
+    public List<BloodRequest> getSpecificRequests(UUID token, String uuid) {
+        if (!permissionsService.hasPermission(token, Permission.ADMIN)) {
+            return null;
+        }
         return bloodRequestService.findSpecific(uuid);
     }
 }
